refactor(grid): move row editng tests to separete file #5664
